* commit '298205cb89ea80887827bf7df9c93cbf4d173129': (28 commits)
  fix comment style
  HAEINSA-31 add line size violation on checkstyle
  HAEINSA-30 apply code formatting
  apply checkstyle with eclipse formatter
  remove trailing whitespaces
  add AvoidNestedBlocks module in checkstyle
  HAEINSA-29 use try-with-resource pattern
  change exception message on chekcSingleRowLock filaed
  HAEINSA-28 update jdk version in pom.xml to 1.7
  remove unnecessary new line
  add forcing unix line separator in checkstyle
  add unmaintainable javadoc tags rules in checkstyle
  add empty block rule in check style
  remove unnecessary paren
  remove consecutive new line
  remove tailing whitespace
  add trailing whitespace rule in checkstyle
  apply naming convention in checkstyle
  normalize checkstyle.xml contents
  do not allow static start imports in checkstyle
  ...

+++ src/main/java/kr/co/vcnc/haeinsa/HaeinsaKeyValue.java
@@ -12,57 +12,58 @@ import com.google.common.base.Objects;
 import com.google.common.collect.ComparisonChain;
 
 /**
- * Modified POJO container of {@link KeyValue} class in HBase.
- * Like {@link KeyValue}, contains only one Key-Value data.
- * 
- * <p>HaeinsaKeyValue contains row, family, qualifier, value information and Type information, but not timestamp.
- * Because haeinsa use timestamp for version control, user cannot manually control timestamp of HaeinsaKeyValue. 
- * Type is same Enum with {@link org.apache.hadoop.hbase.KeyValue.Type}.
- * 
- * <p>HaeinsaKeyValue has public static comparator which can be used in navigableMap.
- * This comparator is ComparisionChain of {@link NullableCompator} wrapped {@link org.apache.hadoop.hbase.util.Bytes#BYTES_COMPARATOR}.
- * The order of comparisonChain is row, family, qualifier, value and type.
- * @author Youngmok Kim
- *
+ * Modified POJO container of {@link KeyValue} class in HBase. Like
+ * {@link KeyValue}, contains only one Key-Value data.
+ * <p>
+ * HaeinsaKeyValue contains row, family, qualifier, value information and Type
+ * information, but not timestamp. Because haeinsa use timestamp for version
+ * control, user cannot manually control timestamp of HaeinsaKeyValue. Type is
+ * same Enum with {@link org.apache.hadoop.hbase.KeyValue.Type}.
+ * <p>
+ * HaeinsaKeyValue has public static comparator which can be used in
+ * navigableMap. This comparator is ComparisionChain of {@link NullableCompator}
+ * wrapped {@link org.apache.hadoop.hbase.util.Bytes#BYTES_COMPARATOR}. The
+ * order of comparisonChain is row, family, qualifier, value and type.
  */
 public class HaeinsaKeyValue {
-	public static final Comparator<HaeinsaKeyValue> COMPARATOR =new Comparator<HaeinsaKeyValue>() {
+	public static final Comparator<HaeinsaKeyValue> COMPARATOR = new Comparator<HaeinsaKeyValue>() {
 		@Override
 		public int compare(HaeinsaKeyValue o1, HaeinsaKeyValue o2) {
 			return ComparisonChain.start()
 					.compare(o1.getRow(), o2.getRow(), new NullableComparator<byte[]>(Bytes.BYTES_COMPARATOR))
 					.compare(o1.getFamily(), o2.getFamily(), new NullableComparator<byte[]>(Bytes.BYTES_COMPARATOR))
 					.compare(o1.getQualifier(), o2.getQualifier(), new NullableComparator<byte[]>(Bytes.BYTES_COMPARATOR))
-					.compare((int)(o2.getType().getCode() & 0xFF), (int)(o1.getType().getCode() & 0xFF))
+					.compare(o2.getType().getCode() & 0xFF, o1.getType().getCode() & 0xFF)
 					.result();
 		}
 	};
-	
+
 	private byte[] row;
 	private byte[] family;
 	private byte[] qualifier;
 	private byte[] value;
 	private Type type;
-	
-	public HaeinsaKeyValue(){
+
+	public HaeinsaKeyValue() {
 	}
-	
-	public HaeinsaKeyValue(KeyValue keyValue){
-		this(keyValue.getRow(), keyValue.getFamily(), keyValue.getQualifier(), keyValue.getValue(), KeyValue.Type.codeToType(keyValue.getType()));
+
+	public HaeinsaKeyValue(KeyValue keyValue) {
+		this(keyValue.getRow(), keyValue.getFamily(), keyValue.getQualifier(), keyValue.getValue(),
+				KeyValue.Type.codeToType(keyValue.getType()));
 	}
-	
-	public HaeinsaKeyValue(byte[] row, byte[] family, byte[] qualifier, byte[] value, Type type){
+
+	public HaeinsaKeyValue(byte[] row, byte[] family, byte[] qualifier, byte[] value, Type type) {
 		this.row = row;
 		this.family = family;
 		this.qualifier = qualifier;
 		this.value = value;
 		this.type = type;
 	}
-	
+
 	public byte[] getRow() {
 		return row;
 	}
-	
+
 	public void setRow(byte[] row) {
 		this.row = row;
 	}
@@ -98,22 +99,22 @@ public class HaeinsaKeyValue {
 	public void setType(Type type) {
 		this.type = type;
 	}
-	
-	public boolean matchingColumn(byte[] family, byte[] qualifier){
+
+	public boolean matchingColumn(byte[] family, byte[] qualifier) {
 		return Bytes.equals(this.family, family) && Bytes.equals(this.qualifier, qualifier);
 	}
-	
+
 	/**
 	 * for debugging
 	 */
 	@Override
-	public String toString(){
+	public String toString() {
 		return Objects.toStringHelper(this.getClass())
 				.add("row", Bytes.toStringBinary(row))
-	            .add("family", Bytes.toStringBinary(family))
-	            .add("qualifier", Bytes.toStringBinary(qualifier))
-	            .add("value", Bytes.toStringBinary(value))
-	            .add("type", type)
-	            .toString();
+				.add("family", Bytes.toStringBinary(family))
+				.add("qualifier", Bytes.toStringBinary(qualifier))
+				.add("value", Bytes.toStringBinary(value))
+				.add("type", type)
+				.toString();
 	}
 }
